disallow trying to remove last user

+++ musidex-daemon/src/application/user_handlers.rs
@@ -2,8 +2,9 @@ use crate::application::handlers::parse_body;
 use crate::domain::entity::{User, UserID};
 use crate::infrastructure::db::Db;
 use crate::infrastructure::router::RequestExt;
+use crate::utils::res_status;
 use anyhow::{Context, Result};
-use hyper::{Body, Request, Response};
+use hyper::{Body, Request, Response, StatusCode};
 use serde::Deserialize;
 
 #[derive(Deserialize)]
@@ -28,6 +29,13 @@ pub async fn delete(req: Request<Body>) -> Result<Response<Body>> {
 
     let db = req.state::<Db>();
     let c = db.get().await;
+
+    if User::n_users(&c)? == 1 {
+        return Ok(Response::builder()
+            .status(StatusCode::UNAUTHORIZED)
+            .body(Body::from("cannot remove last user"))
+            .unwrap());
+    }
 
     User::delete(&c, UserID(id))?;
